refactor: move hooks to dedicated files
